[core] Move `Device::configure_surface` to `Surface::configure`
