+++ player/src/bin/play.rs
@@ -186,7 +186,7 @@ fn main() {
                                     self.resize_config = Some(config);
                                     break;
                                 } else {
-                                    let error = self.device.configure_surface(surface, &config);
+                                    let error = surface.configure(self.device, &config);
                                     self.configured_surface_id = Some(surface_id);
                                     if let Some(e) = error {
                                         panic!("{e:?}");
@@ -231,7 +231,7 @@ fn main() {
                     },
                     WindowEvent::Resized(_) => {
                         if let Some(config) = self.resize_config.take() {
-                            let error = self.device.configure_surface(surface, &config);
+                            let error = surface.configure(self.device, &config);
                             if let Some(e) = error {
                                 panic!("{e:?}");
                             }
